Use Glide for async bitmap loading.
This is what Camera2 code uses.

The thumbnail generated by Glide is not updated
when user rotate the image manually.
This is because exif is rewrite using nio mmap,
however there is no nio unmap available in java.
Even though the file channel is closed, the rotation
modification is still in buffer, when Glide tries
to dtermine the orientation of the image, it still
reads the old data.

Change-Id: I9e76ab40d23213c7a74c26779e345207d5fd7bae